Update int to double

+++ database/migrations/2020_11_11_014643_create_posts_table.php
@@ -19,8 +19,8 @@ class CreatePostsTable extends Migration
           $table->text('body');
           $table->bigInteger('user_id')->unsigned();
           $table->bigInteger('forum_id')->unsigned();
-          $table->int('s_score');
-          $table->int('s_magnitude');
+          $table->double('s_score');
+          $table->double('s_magnitude');
 
           $table->timestamps();
